requirements.txt
numpy
tensorflow-gpu
opencv-python
scikit-image
sklearn
progress
Keras
ipython
segmentation-models
pytz
tensorboard
pillow
pandas
google-cloud-storage
pyyaml
jupyter
crcmod
gitpython
matplotlib
ipykernel
scipy
psutil
GPUtil